hosting: share native plugin editor bridge
